1. Fix memory pool leaks in progress (note it is still leaking..) 2. Add lexeme value to lexer. Parser now can extract the attributes associated with the token. like the string value of an identifier. 3. Misc updates to error and hcc..

+++ trunk/clexer.h
@@ -60,7 +60,7 @@ typedef union lexeme_value
 {
 	int i;
 	float f;
-	double d;
+	long double d;
 	char* s;
 } t_lexeme_value;
